agent-stack/backend/agent/requirements.txt
strands-agents
strands-agents-tools
uv
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
mcp
nest-asyncio
requests
matplotlib
pandas
seaborn
plotly